fix(catalogs): carry the age rating on anime catalog rows
Anime catalog rows are built by parseAnimeCatalogMeta rather than by getMeta,
so the certification the parser had already resolved stayed on the meta as a
bare `certification` field and never reached `app_extras`, which is where
applyDisplayAgeRatingProjection looks for it. Movie and series rows go through
getMeta and so ship the rating in `links`; anime rows shipped no links at all
and the chip appeared only once the meta request landed behind the row, which
reads as the detail page changing under you.

The value was already correct and already computed - only its placement was
wrong - so this restates it where the projection can find it and projects the
row the way the meta builders do. Titles that are genuinely unrated are left
untouched via isUnratedCertification, so "None" does not become a chip.

The projection and its link-stripping helper move to utils/ageRating, next to
the rest of the rating logic, so the catalog and meta paths share one
implementation instead of duplicating it. It is idempotent: an existing chip is
stripped before the current one is prepended, so a meta projected twice is
unchanged.

Verified against a live config over mal.top_anime, anilist.trending, mal.airing
and mal.seasons: every row now carries either the same rating its meta returns
or no chip when genuinely unrated, and the meta path still emits exactly one
rating link.
