+++ addon/utils/ageRating.ts
@@ -41,6 +41,34 @@ export function hasAgeRatingCap(config: { ageRating?: unknown } | null | undefin
   return typeof cap === 'string' && cap !== '' && cap.toLowerCase() !== 'none';
 }
 
+export function stripCertificationLinks(links: any[], certification: string): any[] {
+  if (!Array.isArray(links) || !certification) return links;
+  return links.filter((link: any) => !(link?.name === certification && link?.category === 'Genres'));
+}
+
+/**
+ * Clients read an age rating off the first of `meta.links`, so the certification has to
+ * be restated there. Idempotent: any chip already present is stripped before the current
+ * one is prepended, which lets a meta be projected more than once without stacking.
+ */
+export function applyDisplayAgeRatingProjection(meta: any, config: any): any {
+  const certification = meta?.app_extras?.certification;
+  if (!certification) return meta;
+  const displayCert = meta?.app_extras?.certificationLocal || certification;
+  const links = Array.isArray(meta.links) ? stripCertificationLinks(meta.links, certification).filter((l: any) => !(l?.name === displayCert && l?.category === 'Genres')) : [];
+  if (config.displayAgeRating) {
+    const imdbId = meta.id?.match(/^tt\d+/)?.[0] || meta.imdb_id || meta._imdbId;
+    const tmdbPath = meta.type === 'series' ? 'tv' : 'movie';
+    const url = imdbId
+      ? `https://www.imdb.com/title/${imdbId}/parentalguide/`
+      : `https://www.themoviedb.org/${tmdbPath}/${meta.id}`;
+    meta.links = [{ name: displayCert, category: 'Genres', url }, ...links];
+  } else if (Array.isArray(meta.links)) {
+    meta.links = links;
+  }
+  return meta;
+}
+
 /**
  * Catalog rows carry no certification at all, so treating unknown as blocked empties
  * them outright. Opting out is the strict reading and stays available.
